[WIP]: move "Follow Me" out of navigator to flight_mode_manager
