+++ src/modules/flight_mode_manager/tasks/AutoFollowMe/follow_target_params.c
@@ -1,6 +1,6 @@
 /****************************************************************************
  *
- *   Copyright (c) 2016 PX4 Development Team. All rights reserved.
+ *   Copyright (c) 2016-2021 PX4 Development Team. All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions
@@ -30,18 +30,6 @@
  * POSSIBILITY OF SUCH DAMAGE.
  *
  ****************************************************************************/
-
-/**
- * @file follow_target_params.c
- *
- * Parameters for follow target mode
- *
- * @author Jimmy Johnson <catch22@fastmail.net>
- */
-
-/*
- * Follow target parameters
- */
 
 /**
  * Minimum follow target altitude
