+++ src/modules/navigator/CMakeLists.txt
@@ -50,7 +50,6 @@ px4_add_module(
 		geofence.cpp
 		enginefailure.cpp
 		gpsfailure.cpp
-		follow_target.cpp
 	DEPENDS
 		git_ecl
 		ecl_geo
