+++ src/modules/navigator/navigation.h
@@ -64,7 +64,6 @@ enum NAV_CMD {
 	NAV_CMD_LAND = 21,
 	NAV_CMD_TAKEOFF = 22,
 	NAV_CMD_LOITER_TO_ALT = 31,
-	NAV_CMD_DO_FOLLOW_REPOSITION = 33,
 	NAV_CMD_VTOL_TAKEOFF = 84,
 	NAV_CMD_VTOL_LAND = 85,
 	NAV_CMD_DELAY = 93,
