+++ src/modules/navigator/navigator_main.cpp
@@ -83,8 +83,7 @@ Navigator::Navigator() :
 	_precland(this),
 	_rtl(this),
 	_engineFailure(this),
-	_gpsFailure(this),
-	_follow_target(this)
+	_gpsFailure(this)
 {
 	/* Create a list of our possible navigation types */
 	_navigation_mode_array[0] = &_mission;
@@ -95,7 +94,6 @@ Navigator::Navigator() :
 	_navigation_mode_array[5] = &_takeoff;
 	_navigation_mode_array[6] = &_land;
 	_navigation_mode_array[7] = &_precland;
-	_navigation_mode_array[8] = &_follow_target;
 
 	_handle_back_trans_dec_mss = param_find("VT_B_DEC_MSS");
 	_handle_reverse_delay = param_find("VT_B_REV_DEL");
@@ -629,11 +627,6 @@ Navigator::run()
 			navigation_mode_new = &_gpsFailure;
 			break;
 
-		case vehicle_status_s::NAVIGATION_STATE_AUTO_FOLLOW_TARGET:
-			_pos_sp_triplet_published_invalid_once = false;
-			navigation_mode_new = &_follow_target;
-			break;
-
 		case vehicle_status_s::NAVIGATION_STATE_MANUAL:
 		case vehicle_status_s::NAVIGATION_STATE_ACRO:
 		case vehicle_status_s::NAVIGATION_STATE_ALTCTL:
@@ -803,7 +796,6 @@ void Navigator::geofence_breach_check(bool &have_geofence_position_data)
 					rep->current.alt = loiter_altitude_amsl;
 					rep->current.valid = true;
 					rep->current.loiter_radius = get_loiter_radius();
-					rep->current.alt_valid = true;
 					rep->current.type = position_setpoint_s::SETPOINT_TYPE_LOITER;
 					rep->current.loiter_direction = 1;
 					rep->current.cruising_throttle = get_cruising_throttle();
